Added_Smart_Enemy
added smart enemy and changed all x and y from int to floats.

+++ GameOne/src/com/tutorial/main/FastEnemy.java
@@ -8,7 +8,7 @@ public class FastEnemy extends GameObject{
 
     private Handler handler;
     
-    public FastEnemy(int x, int y, ID id, Handler handler) {
+    public FastEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
         velX = 8;
         velY = 8;
@@ -16,7 +16,7 @@ public class FastEnemy extends GameObject{
     }
     
     public Rectangle getBounds() {
-       return new Rectangle(x,y,16,16);
+       return new Rectangle((int)x,(int)y,16,16);
     }
 
     public void tick() {
@@ -31,6 +31,6 @@ public class FastEnemy extends GameObject{
 
     public void render(Graphics g) { 
         g.setColor(Color.cyan);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int)x, (int)y, 16, 16);
     } 
 }
