+++ GameOne/src/com/tutorial/main/HorizontalEnemy.java
@@ -8,7 +8,7 @@ public class HorizontalEnemy extends GameObject{
 
     private Handler handler;
     
-    public HorizontalEnemy(int x, int y, ID id, Handler handler) {
+    public HorizontalEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
         velX = 9;
         velY = 2;
@@ -16,7 +16,7 @@ public class HorizontalEnemy extends GameObject{
     }
     
     public Rectangle getBounds() {
-       return new Rectangle(x,y,16,16);
+       return new Rectangle((int)x,(int)y,16,16);
     }
 
     public void tick() {
@@ -31,6 +31,6 @@ public class HorizontalEnemy extends GameObject{
 
     public void render(Graphics g) { 
         g.setColor(Color.cyan);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int)x, (int)y, 16, 16);
     } 
 }
